% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmpoly.R
\name{gmpoly2mvp}
\alias{gmpoly2mvp}
\title{Conversion to 'mvp' polynomial}
\usage{
gmpoly2mvp(pol)
}
\arguments{
\item{pol}{a \code{\link{gmpoly}} object}
}
\value{
A \code{\link[mvp:mvp-package]{mvp}} object.
}
\description{
Convert a \code{\link{gmpoly}} polynomial to a 
  \code{\link[mvp:mvp-package]{mvp}} polynomial.
}
\examples{
library(gmpoly)
pol <- gmpoly("5/2 x^(2,2,3) + 3 x^(1,0,1)")
gmpoly2mvp(pol)
}
