% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{gmpolyConstant}
\alias{gmpolyConstant}
\title{Constant multivariate polynomial}
\usage{
gmpolyConstant(m, value)
}
\arguments{
\item{m}{number of variables, a strictly positive integer}

\item{value}{the constant value of the polynomial; the 
\code{\link[gmp]{as.bigq}} function is applied to this argument, so it can 
be e.g. an integer or a character string such as \code{"2/3"} (avoid 
decimal numbers)}
}
\value{
A \code{\link{gmpoly}} object.
}
\description{
Constructs a constant multivariate polynomial.
}
\examples{
library(gmpoly)
gmpolyConstant(3, "2/3")
}
