% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow.R
\name{gmpolyGrow}
\alias{gmpolyGrow}
\title{Grow polynomial}
\usage{
gmpolyGrow(pol)
}
\arguments{
\item{pol}{a \code{\link{gmpoly}} object}
}
\value{
The "same" multivariate polynomial as \code{pol}, except that it 
  has an additional variable.
}
\description{
Grow a multivariate polynomial by including a new variable.
}
\examples{
library(gmpoly)
pol <- gmpoly("3 x^(1,2) - 1/7 x^(5,3)")
gmpolyGrow(pol)
}
