% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Ops.gmpoly}
\alias{Ops.gmpoly}
\title{Arithmetic operators for multivariate polynomials}
\usage{
\method{Ops}{gmpoly}(e1, e2 = NULL)
}
\arguments{
\item{e1, e2}{for an unary operator, only \code{e1} must be given, a 
\code{\link{gmpoly}} object; for a binary operator, at least one of 
\code{e1} and \code{e2} must be a \code{\link{gmpoly}} object, and the 
other must a \code{\link{gmpoly}} object as well or a scalar; the power
operator (\code{^}) is an exception: one can only raise a 
\code{\link{gmpoly}} object to a positive integer power}
}
\value{
A \code{\link{gmpoly}} object.
}
\description{
Arithmetic operators for multivariate polynomials
}
\examples{
library(gmpoly)
}
