% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmpoly.R
\name{print.gmpoly}
\alias{print.gmpoly}
\title{Print a multivariate polynomial}
\usage{
\method{print}{gmpoly}(x, ...)
}
\arguments{
\item{x}{a \code{\link{gmpoly}} object}

\item{...}{ignored}
}
\value{
No value, just prints the polynomial.
}
\description{
Print a multivariate polynomial of class \code{gmpoly}.
}
